refactor(organisms): merge SfLocator
