--- packages/vue/src/components/organisms/SfStoreLocator/SfStoreLocator.html
@@ -1,35 +1,0 @@
-<div class="sf-store-locator">
-  <div class="sf-store-locator__title" tabindex="0">
-    Found stores: <span>{{stores.length}}</span>
-  </div>
-  <div class="sf-store-locator__wrapper">
-    <div class="sf-store-locator__list">
-      <!-- @slot Use this slot to show stores cards -->
-      <slot :center-on="centerOn" :register-store="registerStore" :remove-store="removeStore" :user-position="userPosition" :calculate-distance="getGeoDistance"></slot>
-    </div>
-    <div class="sf-store-locator__map-wrapper">
-      <l-map class="sf-store-locator__map-wrapper-map" v-if="loaded" ref="map" :options="computedMapOptions" :center.sync="internalCenter" :zoom.sync="internalZoom" :max-zoom="maxZoom"
-        @ready="onMapReady" @locationfound="locationFound" @locationerror="locationError">
-        <l-tilelayer v-bind="tileLayerOptions" :url="tileServerUrl" :attribution="tileServerAttribution"  />
-        <l-control-zoom position="topleft" />
-        <l-control position="topleft" class="leaflet-bar leaflet-control leaflet-control-zoom-in sf-store-locator__map-wrapper-map-my-location-btn">
-          <a title="center on user position" role="button" aria-label="center on user position" href="#" @click.prevent="locateUser">
-            <sf-icon icon="home" />
-          </a>
-        </l-control>
-        <l-marker v-for="(store, index) in stores" :lat-lng="store.latlng" :key="index" v-bind="markerOptions">
-          <l-icon :icon-size="markerIconSize" :icon-anchor="markerIconAnchor">
-            <!-- @slot Use this slot to change the icon of the stores, remember to update `markerIconSize` and `markerIconAnchor` accordingly-->
-              <slot name="marker-icon" >
-                <sf-icon :aria-label="`${store.name} located at ${store.address}`" icon="marker" />
-              </slot>
-          </l-icon>
-        </l-marker>
-      </l-map>
-      <!-- @slot Use this slot to customise the loading indicator while the map librry loads -->
-      <slot v-if="!mapReady" name="map-loading" class="sf-store-locator-map-wrapper-loading">
-        <sf-loader />
-      </slot>
-    </div>
-  </div>
-</div>
